+++ lib/redmine_drawio/helpers/textile_helper.rb
@@ -13,7 +13,7 @@ if Rails::VERSION::STRING < '5.0.0'
                 # that patches the jsToolBar adding the new buttons.
                 # After that, all editors in the page will get the new buttons.
                 content_for :header_tags do
-                    javascript_tag 'if(typeof(Drawio) !== "undefined") Drawio.initToolbar();'
+                    javascript_tag 'if(typeof(Drawio) !== "undefined" && typeof Drawio.initToolbar === "function") Drawio.initToolbar();'
                 end
                 @heads_for_wiki_formatter_with_drawio_included = true
             end
@@ -38,7 +38,7 @@ else
                         # that patches the jsToolBar adding the new buttons.
                         # After that, all editors in the page will get the new buttons.
                         content_for :header_tags do
-                            javascript_tag 'if(typeof(Drawio) !== "undefined") Drawio.initToolbar();'
+                            javascript_tag 'if(typeof(Drawio) !== "undefined" && typeof Drawio.initToolbar === "function") Drawio.initToolbar();'
                         end
                         @heads_for_wiki_formatter_with_drawio_included = true
                     end